api/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
tensorflow==2.20.0
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
python-multipart==0.0.9